Moved private functions out of controller into helper mixin. Cleaned up code a bit.
